+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Upcoming
+* Export `caseofEq` and `otherwise` from module entry point
+
 ## 0.9.3 - 2016-11-11
 ### Added
 - `sum-types/caseof-eq` for matching for equality.